Sub-package for CKKS bootstrapping with improved parameters and new functionalities (#139)
cf. CHANGELOG.md

+++ ring/ring.go
@@ -145,7 +145,7 @@ func (r *Ring) genNTTParams() error {
 
 		for i := 0; i < j; i++ {
 
-			r.RescaleParams[j-1][i] = MForm(r.Modulus[i]-ModExp(r.Modulus[j], int(r.Modulus[i]-2), r.Modulus[i]), r.Modulus[i], r.BredParams[i])
+			r.RescaleParams[j-1][i] = MForm(r.Modulus[i]-ModExp(r.Modulus[j], r.Modulus[i]-2, r.Modulus[i]), r.Modulus[i], r.BredParams[i])
 		}
 	}
 
@@ -160,7 +160,7 @@ func (r *Ring) genNTTParams() error {
 	for i, qi := range r.Modulus {
 
 		// 1.1 Compute N^(-1) mod Q in Montgomery form
-		r.NttNInv[i] = MForm(ModExp(uint64(r.N), int(qi-2), qi), qi, r.BredParams[i])
+		r.NttNInv[i] = MForm(ModExp(uint64(r.N), qi-2, qi), qi, r.BredParams[i])
 
 		// 1.2 Compute Psi and PsiInv in Montgomery form
 		r.NttPsi[i] = make([]uint64, r.N)
@@ -169,10 +169,10 @@ func (r *Ring) genNTTParams() error {
 		// Finds a 2N-th primitive Root
 		g := primitiveRoot(qi)
 
-		_2n := r.N << 1
+		_2n := uint64(r.N << 1)
 
-		power := (int(qi) - 1) / _2n
-		powerInv := (int(qi) - 1) - power
+		power := (qi - 1) / _2n
+		powerInv := (qi - 1) - power
 
 		// Computes Psi and PsiInv in Montgomery form
 		PsiMont := MForm(ModExp(g, power, qi), qi, r.BredParams[i])
